LPS-68848 Move packages
