+++ portal-web/docroot/WEB-INF/slim-runtime-web.xml
@@ -21,7 +21,7 @@
 	</listener>
 	<servlet>
 		<servlet-name>Slim Runtime Servlet</servlet-name>
-		<servlet-class>com.liferay.slim.runtime.internal.SlimRuntimeServlet</servlet-class>
+		<servlet-class>com.liferay.slim.runtime.internal.servlet.SlimRuntimeServlet</servlet-class>
 		<load-on-startup>0</load-on-startup>
 		<async-supported>true</async-supported>
 	</servlet>
